+++ yampt.translator/source/editor/operation_executor.cpp
@@ -18,10 +18,10 @@ std::string operation_executor_t::make_output_path(const std::string & source_pa
 {
 	const auto info = QFileInfo(QString::fromStdString(source_path));
 	const auto base_name = info.completeBaseName().toStdString();
-	return make_output_path(base_name, ext, true);
+	return make_output_path_from_name(base_name, ext);
 }
 
-std::string operation_executor_t::make_output_path(const std::string & base_name, const std::string & ext, bool) const
+std::string operation_executor_t::make_output_path_from_name(const std::string & base_name, const std::string & ext) const
 {
 	const auto dir = get_output_dir();
 
@@ -117,7 +117,7 @@ operation_executor_t::result_t operation_executor_t::make_base(
 
 	output_name += "_BASE_" + fl + "-" + nl;
 
-	const auto output_path = make_output_path(output_name, "json", true);
+	const auto output_path = make_output_path_from_name(output_name, "json");
 	dict_writer_t::write(creator.get_dict(), output_path);
 
 	return { !app_logger_t::has_error(), app_logger_t::get_log(), output_path };